run the conv_uint8 script

+++ src/asf/sam0/utils/cmsis/samd20/include/component/pac.h
@@ -57,7 +57,7 @@
 #if !(defined(__ASSEMBLY__) || defined(__IAR_SYSTEMS_ASM__))
 typedef union {
   struct {
-    uint32_t :1;               /*!< bit:      0  Reserved                           */
+    unsigned long int :1;               /*!< bit:      0  Reserved                           */
     uint32_t WP:31;            /*!< bit:  1..31  Write Protection Clear             */
   } bit;                       /*!< Structure used for bit  access                  */
   uint32_t reg;                /*!< Type      used for register access              */
@@ -76,7 +76,7 @@ typedef union {
 #if !(defined(__ASSEMBLY__) || defined(__IAR_SYSTEMS_ASM__))
 typedef union {
   struct {
-    uint32_t :1;               /*!< bit:      0  Reserved                           */
+    unsigned long int :1;               /*!< bit:      0  Reserved                           */
     uint32_t WP:31;            /*!< bit:  1..31  Write Protection Set               */
   } bit;                       /*!< Structure used for bit  access                  */
   uint32_t reg;                /*!< Type      used for register access              */
